[IMP] product: Remove reinvoice cost in product. - Remove "Reinvoice Cost" option in Product Form View. now everything managed through "Invoice on Delivery quantity". - Leave the "At Cost/At Sale Price" option, make it appear only when the product is "invoice on delivery" and when the check box "can be expensed" is checked. - Tool tip on "can be expensed". - In SO Confirm create "Analytic Accounts" automatically when there is a product with "can be expensed" is checked. - In SO, ATTRS on analytic account field = not read only if not set. ========= Technical ========= AccountAnalyticLine class code factorize in sale, sale_expense, sale_timesheet: - Rename the method name '_compute_analytic' to '_compute_analytic_qty'. - Rename the method name '_get_sale_order_line' to '_process_expense_sale_order_line' and moved it from 'sale' to 'sale_expense' then refactored it also. Because it was called only from sale_expense. - Rename the method name '_get_sale_order_line' to '_process_timesheet_sale_order_line' then refactored it in 'sale_ timesheet'. - Test case 'test_cost_invoicing' moved it from 'sale' to 'sale_expense' test. Because this test case is used when 'sale_expense' is installed. - Field 'related_project_id' is related to 'project_id', this field is visible when the Sale Order is in 'sale' state and if it's not set then it should be editable and if is set then it should be read-only. - Removed _get_analytic_invoice_policy and _get_analytic_track_service methods.      Reason: Create the analytic account explicitly when needed.
