+++ addons/hr_expense/models/product_template.py
@@ -7,4 +7,5 @@ from openerp import fields, models
 class ProductTemplate(models.Model):
     _inherit = "product.template"
 
-    can_be_expensed = fields.Boolean(help="Specify whether the product can be selected in an HR expense.")
+    can_be_expensed = fields.Boolean(help='Specify whether the product can be selected in an HR expense.\n'
+                                          'Depending on the modules installed, this will allow you to choose an expense invoice policy if invoice policy is delivered quantities.')
