+++ addons/sale/sale_product_demo.xml
@@ -145,10 +145,6 @@
             <field name="invoice_policy">delivery</field>
         </record>
 
-        <record id="product.service_cost_01" model="product.product">
-            <field name="invoice_policy">cost</field>
-        </record>
-
         <record id="product.service_order_01" model="product.product">
             <field name="invoice_policy">order</field>
         </record>
