+++ addons/sale_expense/models/__init__.py
@@ -2,3 +2,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 import analytic
 import product
+import sale_order
